requirements.txt
opencv-contrib-python
pandas
numpy
scikit-learn
streamlit
pillow
face-recognition
